+++ ChangeLog.md
@@ -13,6 +13,7 @@ v1.0.2
 - Floating Window with tabs (Console/Patch/Camomile)
 - Display patch description
 - Add support for compatibility versions (plugin's version of the patch <= plugin's version used)
+- Add support for block size inferior to 64 samples (implies delay)
 
 v1.0.1
 - Fix thread concurrency issue that occurred when selecting a program (#77).